backend/requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
pydantic==2.5.0
motor==3.3.2
Pillow==10.1.0
cryptography==45.0.5
redis==6.2.0
sendgrid==6.12.4
twilio==9.6.5
authlib==1.6.0
slowapi==0.1.9
PyYAML==6.0.2
tzdata